refactor(farms): continuation of farms refactor
