+++ src/features/farm/FarmList.tsx
@@ -10,7 +10,6 @@ import useSortableData from '../../hooks/useSortableData'
 const FarmList = ({ farms, term }) => {
   const { items, requestSort, sortConfig } = useSortableData(farms)
   const { i18n } = useLingui()
-
   return items ? (
     <>
       <div className="grid grid-cols-4 text-base font-bold text-primary">
@@ -48,9 +47,10 @@ const FarmList = ({ farms, term }) => {
               (sortConfig.direction === 'descending' && <ChevronDownIcon width={12} height={12} />))}
         </div>
       </div>
+
       <div className="flex-col space-y-4">
-        {items.map((farm) => {
-          return <FarmListItem key={`${farm.chef}_${farm.id}`} farm={farm} />
+        {items.map((farm, i) => {
+          return <FarmListItem key={i} farm={farm} />
         })}
       </div>
     </>
