+++ src/services/graph/fetchers/index.ts
@@ -1,8 +1,11 @@
 import {
   ethPriceQuery,
   liquidityPositionSubsetQuery,
+  masterChefV1PairAddressesQuery,
   masterChefV1SushiPerBlockQuery,
   masterChefV1TotalAllocPointQuery,
+  masterChefV2PairAddressesQuery,
+  miniChefPairAddressesQuery,
   miniChefPoolsQuery,
   pairSubsetQuery,
   pairsQuery,
@@ -51,7 +54,15 @@ export const getMasterChefV1SushiPerBlock = async () => {
 }
 
 export const getMasterChefV1Farms = async () => {
-  const { pools } = await masterChefV1(poolsQuery)
+  const { pools } = await request(
+    `https://api.thegraph.com/subgraphs/name/${MASTERCHEF_V1[ChainId.MAINNET]}`,
+    poolsQuery
+  )
+  return pools
+}
+
+export const getMasterChefV1PairAddreses = async () => {
+  const { pools } = await masterChefV1(masterChefV1PairAddressesQuery)
   return pools
 }
 
@@ -60,8 +71,18 @@ export const getMasterChefV2Farms = async () => {
   return pools
 }
 
+export const getMasterChefV2PairAddreses = async () => {
+  const { pools } = await masterChefV2(masterChefV2PairAddressesQuery)
+  return pools
+}
+
 export const getMiniChefFarms = async (chainId = ChainId.MAINNET) => {
   const { pools } = await miniChef(miniChefPoolsQuery, chainId)
+  return pools
+}
+
+export const getMiniChefPairAddreses = async (chainId = ChainId.MAINNET) => {
+  const { pools } = await miniChef(miniChefPairAddressesQuery, chainId)
   return pools
 }
