+++ src/services/graph/hooks/bentobox.ts
@@ -1,18 +1,27 @@
+import { useEffect, useMemo } from 'react'
 import useSWR, { SWRConfiguration } from 'swr'
 
-import { getLendingPairSubset } from '../fetchers/bentobox'
+import { getKashiPairs } from '../fetchers/bentobox'
 import { useActiveWeb3React } from '../../../hooks'
 
-export function useLendingPairSubset(pairAddresses, swrConfig: SWRConfiguration = undefined) {
+export function useKashiPairs(pairAddresses, swrConfig: SWRConfiguration = undefined) {
   const { chainId } = useActiveWeb3React()
-
   const shouldFetch = chainId && pairAddresses && pairAddresses.length
 
-  const res = useSWR(
-    shouldFetch ? ['lendingPairSubset', chainId, pairAddresses] : null,
-    (_, chainId, pairAddresses) => getLendingPairSubset(chainId, { pairAddresses }),
+  // useEffect(() => {
+  //   console.log('debug', { shouldFetch, chainId, pairAddresses })
+  // }, [shouldFetch, chainId, pairAddresses])
+
+  const { data } = useSWR(
+    shouldFetch ? () => ['kashiPairs', chainId, pairAddresses] : null,
+    (_, chainId, pairAddresses) => getKashiPairs(chainId, { pairAddresses }),
     swrConfig
   )
 
-  return res
+  // console.log('kashi pairs data', { data })
+
+  return useMemo(() => {
+    if (!data) return []
+    return data
+  }, [data])
 }
