+++ src/services/graph/hooks/blocks.ts
@@ -6,19 +6,19 @@ import useSWR from 'swr'
 export function useOneDayBlock(swrConfig = undefined) {
   const { chainId } = useActiveWeb3React()
 
-  const res = useSWR(chainId ? ['oneDayBlock', chainId] : null, (_, chainId) => getOneDayBlock(chainId), swrConfig)
+  const { data } = useSWR(chainId ? ['oneDayBlock', chainId] : null, (_, chainId) => getOneDayBlock(chainId), swrConfig)
 
-  return res
+  return data
 }
 
 export function useAverageBlockTime(swrConfig = undefined) {
   const { chainId } = useActiveWeb3React()
 
-  const res = useSWR(
+  const { data } = useSWR(
     chainId ? ['averageBlockTime', chainId] : null,
     (_, chainId) => getAverageBlockTime(chainId),
     swrConfig
   )
 
-  return res
+  return data
 }
