+++ src/services/graph/queries/masterchef-v2.ts
@@ -19,3 +19,19 @@ export const poolsV2Query = gql`
     }
   }
 `
+
+export const masterChefV2PairAddressesQuery = gql`
+  query masterChefV2PairAddresses(
+    $first: Int! = 1000
+    $skip: Int! = 0
+    $orderBy: String! = "timestamp"
+    $orderDirection: String! = "desc"
+  ) {
+    pools(first: $first, skip: $skip, orderBy: $orderBy, orderDirection: $orderDirection) {
+      id
+      pair {
+        id
+      }
+    }
+  }
+`
