+++ src/services/graph/queries/masterchef.ts
@@ -24,6 +24,22 @@ export const poolsQuery = gql`
   }
 `
 
+export const masterChefV1PairAddressesQuery = gql`
+  query masterChefV1PairAddresses(
+    $first: Int! = 1000
+    $skip: Int! = 0
+    $orderBy: String! = "timestamp"
+    $orderDirection: String! = "desc"
+  ) {
+    pools(first: $first, skip: $skip, orderBy: $orderBy, orderDirection: $orderDirection) {
+      id
+      pair {
+        id
+      }
+    }
+  }
+`
+
 export const masterChefV1TotalAllocPointQuery = gql`
   query masterChefV1TotalAllocPoint($id: String! = "0xc2edad668740f1aa35e4d8f227fb8e17dca888cd") {
     masterChef(id: $id) {
