+++ src/services/graph/queries/minichef.ts
@@ -28,3 +28,19 @@ export const miniChefPoolsQuery = gql`
     }
   }
 `
+
+export const miniChefPairAddressesQuery = gql`
+  query miniChefPairAddresses(
+    $first: Int! = 1000
+    $skip: Int! = 0
+    $orderBy: String! = "timestamp"
+    $orderDirection: String! = "desc"
+  ) {
+    pools(first: $first, skip: $skip, orderBy: $orderBy, orderDirection: $orderDirection) {
+      id
+      pair {
+        id
+      }
+    }
+  }
+`
